Update anonymous-users-in-meetings.md
Added step to the Meeting Policy because you have to choose a policy. But the section on Meeting Settings vs Global Settings for anonymous users seems really backwards. Meeting settings can't be applied to people. Meeting policies can. So it makes more sense that the meeting SETTINGS are the org-wide, and the meeting policies can be done per-organizer. Except that there's still the global settings. 

I tried getting Co-Pilot to explain this and it just made him contradict himself over and over.